fix some harmless PVS warnings
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@4587 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/sound/m114s.c
@@ -24,7 +24,8 @@
  *
  *   The mixer can optionally mix every (internal) channel into a separate stream for testing (set M114S_OUTPUT_CHANNELS to 16)
  *
- *   TODO: No full/tested support for the M114AF 6Mhz version of the chip (but seems at least to work 'good enough')
+ *   TODO: - No full/tested support for the M114AF 6Mhz version of the chip (but seems at least to work 'good enough')
+ *         - Maybe also low pass filter heavily/some channels only?? (some channels sound too high frequency heavy (clicks, pops), for example on Dakar)
  **********************************************************************************************/
 
 #include "driver.h"
@@ -511,7 +512,7 @@ static void m114s_update(int num,
 
 		/* Update the buffer & Ensure we don't clip */
 #if M114S_OUTPUT_CHANNELS == 1
-		accum /= 4;
+		accum /= (INT32)4;
 		*buffer++ = (accum < -32768) ? -32768 : ((accum > 32767) ? 32767 : accum);
 #else
 		for (c = 0; c < M114S_OUTPUT_CHANNELS; c++)
